Move PodTemplate -> Pod conversion to PodTemplateBuilder
