+++ src/main/java/org/csanchez/jenkins/plugins/kubernetes/PodTemplate.java
@@ -32,6 +32,7 @@ import hudson.model.Label;
 import hudson.model.Node;
 import hudson.model.labels.LabelAtom;
 import hudson.tools.ToolLocationNodeProperty;
+import io.fabric8.kubernetes.api.model.Pod;
 import jenkins.model.Jenkins;
 
 /**
@@ -587,6 +588,15 @@ public class PodTemplate extends AbstractDescribableImpl<PodTemplate> implements
         return this;
     }
 
+    /**
+     * Build a Pod object from a PodTemplate
+     * 
+     * @param slave
+     */
+    public Pod build(KubernetesSlave slave) {
+        return new PodTemplateBuilder(this).build(slave);
+    }
+
     @Extension
     public static class DescriptorImpl extends Descriptor<PodTemplate> {
 
@@ -601,4 +611,5 @@ public class PodTemplate extends AbstractDescribableImpl<PodTemplate> implements
             return DescriptorVisibilityFilter.apply(null, Jenkins.getInstance().getDescriptorList(TemplateEnvVar.class));
         }
     }
+
 }
