+++ src/test/java/org/csanchez/jenkins/plugins/kubernetes/PodTemplateBuilderTest.java
@@ -1,0 +1,32 @@
+package org.csanchez.jenkins.plugins.kubernetes;
+
+import static org.csanchez.jenkins.plugins.kubernetes.PodTemplateBuilder.*;
+import static org.junit.Assert.*;
+
+import org.junit.Test;
+
+import com.google.common.collect.ImmutableList;
+
+public class PodTemplateBuilderTest {
+    @Test
+    public void testParseDockerCommand() {
+        assertNull(parseDockerCommand(""));
+        assertNull(parseDockerCommand(null));
+        assertEquals(ImmutableList.of("bash"), parseDockerCommand("bash"));
+        assertEquals(ImmutableList.of("bash", "-c", "x y"), parseDockerCommand("bash -c \"x y\""));
+        assertEquals(ImmutableList.of("a", "b", "c", "d"), parseDockerCommand("a b c d"));
+    }
+
+    @Test
+    public void testParseLivenessProbe() {
+        assertNull(parseLivenessProbe(""));
+        assertNull(parseLivenessProbe(null));
+        assertEquals(ImmutableList.of("docker", "info"), parseLivenessProbe("docker info"));
+        assertEquals(ImmutableList.of("echo", "I said: 'I am alive'"),
+                parseLivenessProbe("echo \"I said: 'I am alive'\""));
+        assertEquals(ImmutableList.of("docker", "--version"), parseLivenessProbe("docker --version"));
+        assertEquals(ImmutableList.of("curl", "-k", "--silent", "--output=/dev/null", "https://localhost:8080"),
+                parseLivenessProbe("curl -k --silent --output=/dev/null \"https://localhost:8080\""));
+    }
+
+}
